recommendations_system/requirements.txt
google-genai
pypdf
python-dotenv

